+++ tests/_mocks/Items/MasterItem.php
@@ -63,4 +63,9 @@ class MasterItem extends Item
     {
         return $this->morphToMany();
     }
+
+    public function empty()
+    {
+        return $this->hasOne(ChildItem::class);
+    }
 }